refactor(sdk): extract shared env helpers and pin dependencies
loadDotEnv and stripEnvQuotes were byte-identical in all three smoke
tests and maskKey/maskToken in two of them; they move to sdk/lib/env.mjs
with the semantics preserved verbatim — including continuing silently
when .env is absent, which is why node --env-file is not an option here
(it aborts on a missing file, and --env-file-if-exists needs a newer
runtime than the declared engines range).

ai and @ai-sdk/openai-compatible were declared as "latest", so any npm
install could silently jump a major version past the lockfile; both are
now pinned to the versions the lockfile already resolves.
