+++ sdk/test-deepseek.mjs
@@ -1,10 +1,11 @@
 #!/usr/bin/env node
 
-import { existsSync, readFileSync } from "node:fs";
 import { dirname, resolve } from "node:path";
 import { fileURLToPath } from "node:url";
 import { createOpenAICompatible } from "@ai-sdk/openai-compatible";
 import { generateText } from "ai";
+
+import { loadDotEnv, maskKey } from "./lib/env.mjs";
 
 const __dirname = dirname(fileURLToPath(import.meta.url));
 loadDotEnv(resolve(__dirname, ".env"));
@@ -99,36 +100,9 @@ async function checkModelVisibility({ baseURL, apiKey, model }) {
   console.log("");
 }
 
-function loadDotEnv(path) {
-  if (!existsSync(path)) return;
-  const lines = readFileSync(path, "utf8").split(/\r?\n/);
-  for (const line of lines) {
-    const trimmed = line.trim();
-    if (!trimmed || trimmed.startsWith("#")) continue;
-    const eq = trimmed.indexOf("=");
-    if (eq < 0) continue;
-    const key = trimmed.slice(0, eq).trim();
-    const raw = trimmed.slice(eq + 1).trim();
-    if (!key || process.env[key] !== undefined) continue;
-    process.env[key] = stripEnvQuotes(raw);
-  }
-}
-
-function stripEnvQuotes(value) {
-  if ((value.startsWith('"') && value.endsWith('"')) || (value.startsWith("'") && value.endsWith("'"))) {
-    return value.slice(1, -1);
-  }
-  return value;
-}
-
 function normalizeBaseURL(value) {
   const trimmed = value.replace(/\/+$/, "");
   return trimmed.endsWith("/v1") ? trimmed : `${trimmed}/v1`;
-}
-
-function maskKey(value) {
-  if (value.length <= 10) return "***";
-  return `${value.slice(0, 6)}...${value.slice(-4)}`;
 }
 
 function formatAIError(error) {
