+++ sdk/test-security-policy.mjs
@@ -1,8 +1,9 @@
 #!/usr/bin/env node
 
-import { existsSync, readFileSync } from "node:fs";
 import { dirname, resolve } from "node:path";
 import { fileURLToPath } from "node:url";
+
+import { loadDotEnv, maskKey } from "./lib/env.mjs";
 
 const __dirname = dirname(fileURLToPath(import.meta.url));
 loadDotEnv(resolve(__dirname, ".env"));
@@ -39,7 +40,7 @@ console.log("TokenHub security policy smoke test");
 console.log(`Admin URL: ${adminBaseURL}`);
 console.log(`Policy ID: ${policyID}`);
 console.log(`Policy:    ${policyName}`);
-console.log(`Token:     ${maskToken(adminToken)}`);
+console.log(`Token:     ${maskKey(adminToken)}`);
 console.log("");
 
 try {
@@ -144,36 +145,9 @@ function assertPolicy(policy, expectedID) {
   }
 }
 
-function loadDotEnv(path) {
-  if (!existsSync(path)) return;
-  const lines = readFileSync(path, "utf8").split(/\r?\n/);
-  for (const line of lines) {
-    const trimmed = line.trim();
-    if (!trimmed || trimmed.startsWith("#")) continue;
-    const eq = trimmed.indexOf("=");
-    if (eq < 0) continue;
-    const key = trimmed.slice(0, eq).trim();
-    const raw = trimmed.slice(eq + 1).trim();
-    if (!key || process.env[key] !== undefined) continue;
-    process.env[key] = stripEnvQuotes(raw);
-  }
-}
-
-function stripEnvQuotes(value) {
-  if ((value.startsWith('"') && value.endsWith('"')) || (value.startsWith("'") && value.endsWith("'"))) {
-    return value.slice(1, -1);
-  }
-  return value;
-}
-
 function normalizeAdminBaseURL(value) {
   const trimmed = value.replace(/\/+$/, "");
   return trimmed.endsWith("/v1") ? trimmed.slice(0, -3) : trimmed;
-}
-
-function maskToken(value) {
-  if (value.length <= 10) return "***";
-  return `${value.slice(0, 6)}...${value.slice(-4)}`;
 }
 
 function formatError(error) {
